feat: number generator 추가
