+++ heon-playground/baseball-playground/src/main/kotlin/com/gonza/heon/baseballplayground/BaseballConstants.kt
@@ -1,0 +1,7 @@
+package com.gonza.heon.baseballplayground
+
+object BaseballConstants {
+    const val NUMBER_LENGTH = 3
+    const val NUMBER_ORIGIN = 1
+    const val NUMBER_BOUND = 10
+}
